+++ savedsearch-web/src/main/java/oracle/sysman/emSDK/emaas/platform/savedsearch/ws/rest/zdt/rowsEntity/SavedSearchFolderRowEntity.java
@@ -10,6 +10,8 @@
 
 package oracle.sysman.emSDK.emaas.platform.savedsearch.ws.rest.zdt.rowsEntity;
 
+import java.math.BigInteger;
+
 import org.codehaus.jackson.annotate.JsonProperty;
 
 /**
@@ -18,7 +20,7 @@ import org.codehaus.jackson.annotate.JsonProperty;
 public class SavedSearchFolderRowEntity implements RowEntity
 {
 	@JsonProperty("FOLDER_ID")
-	private Long folderId;
+	private BigInteger folderId;
 
 	@JsonProperty("DESCRIPTION")
 	private String description;
@@ -57,13 +59,13 @@ public class SavedSearchFolderRowEntity implements RowEntity
 	private Integer uiHidden;
 
 	@JsonProperty("DELETED")
-	private Long deleted;
+	private BigInteger deleted;
 
 	@JsonProperty("TENANT_ID")
 	private Long tenantId;
 
 	@JsonProperty("PARENT_ID")
-	private Long parentId;
+	private BigInteger parentId;
 
 	@JsonProperty("CREATION_DATE")
 	private String creationDate;
@@ -226,7 +228,7 @@ public class SavedSearchFolderRowEntity implements RowEntity
 	/**
 	 * @return the deleted
 	 */
-	public Long getDeleted()
+	public BigInteger getDeleted()
 	{
 		return deleted;
 	}
@@ -266,7 +268,7 @@ public class SavedSearchFolderRowEntity implements RowEntity
 	/**
 	 * @return the folderId
 	 */
-	public Long getFolderId()
+	public BigInteger getFolderId()
 	{
 		return folderId;
 	}
@@ -322,7 +324,7 @@ public class SavedSearchFolderRowEntity implements RowEntity
 	/**
 	 * @return the parentId
 	 */
-	public Long getParentId()
+	public BigInteger getParentId()
 	{
 		return parentId;
 	}
@@ -391,7 +393,7 @@ public class SavedSearchFolderRowEntity implements RowEntity
 	 * @param deleted
 	 *            the deleted to set
 	 */
-	public void setDeleted(Long deleted)
+	public void setDeleted(BigInteger deleted)
 	{
 		this.deleted = deleted;
 	}
@@ -436,7 +438,7 @@ public class SavedSearchFolderRowEntity implements RowEntity
 	 * @param folderId
 	 *            the folderId to set
 	 */
-	public void setFolderId(Long folderId)
+	public void setFolderId(BigInteger folderId)
 	{
 		this.folderId = folderId;
 	}
@@ -499,7 +501,7 @@ public class SavedSearchFolderRowEntity implements RowEntity
 	 * @param parentId
 	 *            the parentId to set
 	 */
-	public void setParentId(Long parentId)
+	public void setParentId(BigInteger parentId)
 	{
 		this.parentId = parentId;
 	}
